disassemble 분리
