+++ docs/src/pages/docs/api/assemble.ko.mdx
@@ -4,7 +4,6 @@ title: assemble
 
 import { Sandpack } from '@/components/Sandpack';
 
-
 # assemble
 
 인자로 받은 배열에 담긴 한글 문장과 문자를 한글 규칙에 맞게 합성합니다.
@@ -13,10 +12,8 @@ import { Sandpack } from '@/components/Sandpack';
 function assemble(
   // 한글 문자와 문장을 담고 있는 배열
   words: string[]
-): string
+): string;
 ```
-
-
 
 ## 사용해보기
 
@@ -27,10 +24,9 @@ function assemble(
 ```ts index.ts
 import { assemble } from 'es-hangul';
 
-console.log(assemble(['아버지가', ' ', '방ㅇ', 'ㅔ ', '들ㅇ', 'ㅓ갑니다']))
-console.log(assemble(['아버지가', ' ', '방에 ', '들어갑니다']))
-console.log(assemble(['ㅇ', 'ㅏ', 'ㅂ', 'ㅓ', 'ㅈ', 'ㅣ']))
+console.log(assemble(['아버지가', ' ', '방ㅇ', 'ㅔ ', '들ㅇ', 'ㅓ갑니다']));
+console.log(assemble(['아버지가', ' ', '방에 ', '들어갑니다']));
+console.log(assemble(['ㅇ', 'ㅏ', 'ㅂ', 'ㅓ', 'ㅈ', 'ㅣ']));
 ```
 
 </Sandpack>
-
