+++ docs/src/pages/docs/api/disassembleCompleteCharacter.ko.mdx
@@ -1,0 +1,36 @@
+---
+title: disassembleCompleteCharacter
+---
+
+import { Sandpack } from '@/components/Sandpack';
+
+# disassembleCompleteCharacter
+
+완전한 한글 문자열을 초성, 중성, 종성으로 분리합니다.
+
+```typescript
+function disassembleCompleteCharacter(letter: string): ReturnTypeDisassembleCompleteCharacter | undefined;
+```
+
+## Examples
+
+```tsx
+disassembleCompleteCharacter('값'); // { choseong: 'ㄱ', jungseong: 'ㅏ', jongseong: 'ㅂㅅ' }
+disassembleCompleteCharacter('리'); // { choseong: 'ㄹ', jungseong: 'ㅣ', jongseong: '' }
+disassembleCompleteCharacter('빚'); // { choseong: 'ㅂ', jungseong: 'ㅣ', jongseong: 'ㅈ' }
+disassembleCompleteCharacter('박'); // { choseong: 'ㅂ', jungseong: 'ㅏ', jongseong: 'ㄱ' }
+```
+
+## 사용해보기
+
+<br />
+
+<Sandpack>
+
+```ts index.ts
+import { disassembleCompleteCharacter } from 'es-hangul';
+
+console.log(disassembleCompleteCharacter('값'));
+```
+
+</Sandpack>
