+++ src/index.ts
@@ -7,8 +7,9 @@ export { combineCharacter } from './combineCharacter';
 export { combineVowels } from './combineVowels';
 export { convertQwertyToAlphabet, convertQwertyToHangul } from './convertQwertyToAlphabet';
 export { days } from './days';
-export { disassemble, disassembleToGroups } from './disassemble';
+export { disassemble } from './disassemble';
 export { disassembleCompleteCharacter } from './disassembleCompleteCharacter';
+export { disassembleToGroups } from './disassembleToGroups';
 export { getChoseong } from './getChoseong';
 export { hasBatchim } from './hasBatchim';
 export { josa } from './josa';
